add output when invalid activation function

+++ inc/neuronTypes.hpp
@@ -8,8 +8,8 @@ void* operator new(size_t size);
 void operator delete(void* ptr) noexcept;
 void operator delete[](void* ptr) noexcept;
 #endif
-typedef uint8_t weight_t;
-typedef uint8_t in_out_t;
+typedef float weight_t;
+typedef float in_out_t;
 typedef uint16_t count_t;
 
 #endif
